# By Sebastian Hoehna (119) and others
# Via Sebastian Hoehna (26) and others
* 'development' of https://github.com/revbayes/revbayes: (469 commits)
  Adding function for Rev users to get children of given node. Adding fix for clamped unrooted uniform topology trees returning -Inf likelihoods due to internal rooting choices.
  fixing small issue with TreeTrace burnin
  fixing truncated normal travis CI fail
  Revert "allow pdf to compute density outside its support"
  allow pdf to compute density outside its support
  fixing dynamic tip time slide move weight
  fixing some issues with the truncated normal distribution
  cleaning up and error checking in DPPAllocateAuxGibbsMove
  better fix for DPPAllocateAux
  removing non-existent moves
  implementing vector slide/recenter move
  updating expected test output again, CDBDP likelihood is slightly different...
  adding support for origin age in CDBDP
  deleting redundant StateDependentSpeciationExtinctionProcess
  renaming some member variables
  fixing bugs in tree tip moves
  Fixed merge problems.
  minor fix to backbone topology init
  fixing inverse gamma distribution (now parameterized with SHAPE and RATE, as advertized)
  fix DPPAllocateAux move to avoid NaN likelihoods
  ...

Conflicts:
	src/core/datatypes/trees/Tree.cpp
	src/revlanguage/datatypes/phylogenetics/trees/RlTree.cpp
	src/revlanguage/workspace/RbRegister_Move.cpp

+++ src/core/distributions/phylogenetics/tree/birthdeath/SampledSpeciationBirthDeathProcess.h
@@ -20,7 +20,7 @@ namespace RevBayesCore {
     
     class Clade;
     
-    class SampledSpeciationBirthDeathProcess : public AbstractCharacterHistoryBirthDeathProcess, public MemberObject< RbVector<int> >, public MemberObject<RbVector<RbVector<double> > > {
+    class SampledSpeciationBirthDeathProcess : public AbstractCharacterHistoryBirthDeathProcess, public MemberObject< RbVector<long> >, public MemberObject<RbVector<RbVector<double> > > {
         
     public:
         SampledSpeciationBirthDeathProcess(const TypedDagNode<double> *a, const TypedDagNode<double> *s, const TypedDagNode<double> *e, const TypedDagNode<double> *r, const std::vector<Taxon> &n);                                                                                  //!< Constructor
@@ -30,7 +30,7 @@ namespace RevBayesCore {
         // public member functions
         SampledSpeciationBirthDeathProcess*                 clone(void) const;                                          //!< Create an independent clone
         double                                              computeLnProbability(void);                                 //!< Compute ln prob of current value
-        void                                                executeMethod(const std::string &n, const std::vector<const DagNode*> &args, RbVector<int> &rv) const;     //!< Map the member methods to internal function calls
+        void                                                executeMethod(const std::string &n, const std::vector<const DagNode*> &args, RbVector<long> &rv) const;     //!< Map the member methods to internal function calls
         void                                                executeMethod(const std::string &n, const std::vector<const DagNode*> &args, RbVector<RbVector<double> > &rv) const;     //!< Map the member methods to internal function calls
 
         CharacterHistory&                                   getCharacterHistory(void);                                  //!< Get the character histories
