+++ src/revlanguage/distributions/math/Dist_binomial.cpp
@@ -32,7 +32,7 @@ RevBayesCore::BinomialDistribution* Dist_binomial::createDistribution( void ) co
 {
     
     // get the parameters
-    RevBayesCore::TypedDagNode<int>*    vn = static_cast<const Natural     &>( n->getRevObject() ).getDagNode();
+    RevBayesCore::TypedDagNode<long>*    vn = static_cast<const Natural     &>( n->getRevObject() ).getDagNode();
     RevBayesCore::TypedDagNode<double>* vp = static_cast<const Probability &>( p->getRevObject() ).getDagNode();
     RevBayesCore::BinomialDistribution* d  = new RevBayesCore::BinomialDistribution( vn, vp );
     return d;
