+++ src/revlanguage/moves/vector/Move_SynchronizedVectorFixedSingleElementSlide.cpp
@@ -64,7 +64,7 @@ void Move_SynchronizedVectorFixedSingleElementSlide::constructInternalObject( vo
     
     bool t = static_cast<const RlBoolean &>( tune->getRevObject() ).getValue();
     
-    std::vector<int> e;
+    std::vector<long> e;
     if ( which_element->getRevObject().isType( ModelVector<Natural>::getClassTypeSpec() ) )
     {
         e = static_cast<const ModelVector<Natural> &>( which_element->getRevObject() ).getValue();
